utils: fix integer size to avoid overflow on 32 bits (#1328)

+++ src/flb_utils.c
@@ -337,12 +337,12 @@ int flb_utils_pipe_byte_consume(flb_pipefd_t fd)
     return 0;
 }
 
-ssize_t flb_utils_size_to_bytes(char *size)
+int64_t flb_utils_size_to_bytes(char *size)
 {
     int i;
     int len;
     int plen = 0;
-    size_t val;
+    int64_t val;
     char c;
     char tmp[3] = {0};
     int64_t KB = 1000;
